+++ lib/src/db/utilities/eversource/supplier_rates.dart
@@ -1,5 +1,0 @@
-
-
-
-/// See https://energizect.com/rate-board-residential-standard-service-generation-rates
-/// Each month has customer counts by