+++ test/db/isoexpress/da_lmp_hourly_test.dart
@@ -164,10 +164,9 @@ Future<void> tests(String rootUrl) async {
     });
   });
   group('DAM LMP client tests: ', () {
-    var daLmp =
-        client.DaLmp(http.Client(), iso: Iso.newEngland, rootUrl: rootUrl);
+    var daLmp = client.DaLmp(http.Client(), rootUrl: rootUrl);
     test('get daily peak price between two dates', () async {
-      var data = await daLmp.getDailyLmpBucket(4000, LmpComponent.lmp,
+      var data = await daLmp.getDailyLmpBucket(Iso.newEngland, 4000, LmpComponent.lmp,
           IsoNewEngland.bucket5x16, Date.utc(2017, 1, 1), Date.utc(2017, 1, 5));
       expect(data.length, 3);
       expect(data.toList(), [
@@ -177,14 +176,14 @@ Future<void> tests(String rootUrl) async {
       ]);
     });
     test('get monthly peak price between two dates', () async {
-      var data = await daLmp.getMonthlyLmpBucket(4000, LmpComponent.lmp,
+      var data = await daLmp.getMonthlyLmpBucket(Iso.newEngland,4000, LmpComponent.lmp,
           IsoNewEngland.bucket5x16, Month.utc(2017, 1), Month.utc(2017, 8));
       expect(data.length, 8);
       expect(data.first,
           IntervalTuple(Month(2017, 1, location: location), 42.55883928571426));
     });
     test('get hourly price for 2017-01-01', () async {
-      var data = await daLmp.getHourlyLmp(
+      var data = await daLmp.getHourlyLmp(Iso.newEngland,
           4000, LmpComponent.lmp, Date.utc(2017, 1, 1), Date.utc(2017, 1, 1));
       expect(data.length, 24);
       expect(
@@ -193,7 +192,7 @@ Future<void> tests(String rootUrl) async {
               Hour.beginning(TZDateTime(location, 2017, 1, 1)), 35.12));
     });
     test('get daily prices all nodes', () async {
-      var data = await daLmp.getDailyPricesAllNodes(
+      var data = await daLmp.getDailyPricesAllNodes(Iso.newEngland,
           LmpComponent.lmp, Date.utc(2017, 1, 1), Date.utc(2017, 1, 3));
       expect(data.length, 1136);
       var p321 = data[321]!;
@@ -205,9 +204,9 @@ Future<void> tests(String rootUrl) async {
 Future<void> speedTest(String rootUrl) async {
   var location = getLocation('America/New_York');
   var daLmp =
-      client.DaLmp(http.Client(), iso: Iso.newEngland, rootUrl: rootUrl);
+      client.DaLmp(http.Client(), rootUrl: rootUrl);
 
-  var data = await daLmp.getHourlyLmp(
+  var data = await daLmp.getHourlyLmp(Iso.newEngland,
       4000, LmpComponent.lmp, Date.utc(2017, 1, 1), Date.utc(2017, 1, 1));
 }